% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\docType{data}
\name{des}
\alias{des}
\title{Data dengan nama des dari 12 waktu simulasi}
\format{
data frame dengan 12 baris dan 2 variabel:
\describe{
\item{Waktu.Tiba}{Waktu.Tiba, dalam menit}
\item{Lama.layanan}{Lama.Layanan, dalam menit}
...
}
}
\usage{
des
}
\description{
Sebuah dataset yang berisi dua kolom, Waktu.Tiba dan Lama.Layanan
kedua attribut ini menjadi input waktu tiba entiti dan lamanya aktivitas
}
\keyword{datasets}
