% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpel.R
\name{simpel}
\alias{simpel}
\title{Simple antrians sederhana server tunggal.}
\usage{
simpel(ak = 0.4, lm = 0.45, rd = 3, rsc = 1, rn = 50, ongoing = TRUE)
}
\arguments{
\item{ak}{A number.}

\item{lm}{A number.}

\item{rd}{A number.}

\item{rsc}{A number.}

\item{rn}{A number.}

\item{ongoing}{logical.}
}
\value{
a data frame of \code{get_mon_arrivals}
}
\description{
Simple antrians sederhana server tunggal.
}
\examples{
simpel(ak = 0.5, lm = 0.55)
simpel(rn = 5)
}
